# src/api/api.yaml
openapi: 3.0.2
info:
  title: "Evolutionary Manufacturing Scheduler API"
  version: 1.0.0
servers:
  - url: https://localhost:9999/api
paths:
  /site-data:
    get:
      summary: get all site-datas
      operationId: src.api.api.get_sites_data
      responses:
        '200':
          description: collection of sites data
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    post:
      summary: uploading a json file with site data
      operationId: src.api.api.create_site_data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: byte
                title:
                  type: string
              required:
                - file
      responses:
        '200':
          description: site data object created
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /site-data/{site_data_id}:
    get:
      summary: get site-data of specific id
      operationId: src.api.api.get_site_data_by_id
      parameters:
        - in: path
          name: site_data_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get site-data
      responses:
        '200':
          description: site data
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    delete:
      summary: remove site data
      operationId: src.api.api.delete_site_data
      parameters:
        - in: path
          name: site_data_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get site-data
      responses:
        '200':
          description: site data object deleted
  /problem:
    get:
      summary: get all problems
      operationId: src.api.api.get_problems
      responses:
        '200':
          description: collection of problems
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    post:
      summary: create a new problem from a given site data (id)
      operationId: src.api.api.add_problem
      requestBody:
        description: related site data id
        content:
          application/json:
            schema:
              type: object
              properties:
                site_data_id:
                  type: integer
                  format: int32
                title:
                  type: string
        required: true
      responses:
        '200':
          description: successfully created new problem
      x-codegen-request-body-name: body
  /problems/{problem_id}:
    get:
      summary: get problem by id
      operationId: src.api.api.get_problem_by_id
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: problem object
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    delete:
      summary: remove problem
      operationId: src.api.api.delete_problem
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: problem object deleted
  /problem/{problem_id}/selection:
    put:
      summary: change selection method for specific problem
      operationId: src.api.api.edit_selection_method
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      requestBody:
        description: selection configuration
        content:
          application/json:
            schema:
              type: object
              properties:
                selection_id:
                  type: integer
                  format: int32
                selection_parameters:
                  type: object
              required:
                - selection_id
        required: true
      responses:
        '200':
          description: selection modified successfully
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/crossover:
    put:
      summary: change crossover method for specific problem
      operationId: src.api.api.edit_crossover_method
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      requestBody:
        description: crossover configuration
        content:
          application/json:
            schema:
              type: object
              properties:
                crossover_id:
                  type: integer
                  format: int32
                crossover_parameters:
                  type: object
              required:
                - crossover_id
        required: true
      responses:
        '200':
          description: crossover modified successfully
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/mutation:
    post:
      summary: add new mutation for specific problem
      operationId: src.api.api.add_mutation_method
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      requestBody:
        description: mutation configuration
        content:
          application/json:
            schema:
              type: object
              properties:
                mutation_id:
                  type: integer
                  format: int32
                mutation_parameters:
                  type: object
              required:
                - mutation_id
        required: true
      responses:
        '200':
          description: mutation added successfully
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/mutations/{mutation_id}:
    put:
      summary: change mutation for specific problem
      operationId: src.api.api.edit_mutation_method
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
        - in: path
          name: mutation_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      requestBody:
        description: mutation configuration
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: mutation modified successfully
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    delete:
      summary: remove mutation
      operationId: src.api.api.remove_mutation
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
        - in: path
          name: mutation_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get mutation
      responses:
        '200':
          description: mutation object deleted
  /problem/{problem_id}/population:
    put:
      summary: set population size for EA to run
      operationId: src.api.api.edit_ea_population_size
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      requestBody:
        description: size of population
        content:
          application/json:
            schema:
              type: object
              properties:
                population_size:
                  type: integer
                  format: int32
              required:
                - population_size
        required: true
      responses:
        '200':
          description: population size modified successfully
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/stopping-condition/{cond_id}:
    put:
      summary: set stopping-condition size for EA to run
      operationId: src.api.api.set_stopping_condition
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
        - in: path
          name: cond_id
          schema:
            type: string
          required: true
          description: string ID to get stopping-condition
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bound:
                  oneOf:
                    - type: integer
                    - type: string
      responses:
        '200':
          description: population size modified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cond_id:
                    type: integer
                    format: int32
                  bound:
                    oneOf:
                      - type: integer
                      - type: string
      x-codegen-request-body-name: body
    delete:
      summary: delete stopping condition
      operationId: src.api.api.delete_stopping_condition
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
        - in: path
          name: cond_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get stopping-condition
      responses:
        '200':
          description: stopping-condition deleted
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/play:
    post:
      summary: start ea for this problem
      operationId: src.api.api.start_ea
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: algorithm started
      x-codegen-request-body-name: body
  /problem/{problem_id}/pause:
    post:
      summary: pause ea for this problem
      operationId: src.api.api.pause_ea
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: algorithm paused
      x-codegen-request-body-name: body
  /problem/{problem_id}/resume:
    post:
      summary: resume ea for this problem
      operationId: src.api.api.resume_ea
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: algorithm resumed
      x-codegen-request-body-name: body
  /problem/{problem_id}/stop:
    post:
      summary: stop ea for this problem
      operationId: src.api.api.stop_ea
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: algorithm stopped
      x-codegen-request-body-name: body
  /problem/{problem_id}/progress:
    get:
      summary: get progress according to stopping-conditions for this problem
      operationId: src.api.api.get_ea_progress
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: get all stopping-condition progress data
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/solution:
    get:
      summary: get current best solution
      operationId: src.api.api.get_ea_best_solution
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: get best solution
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
    put:
      summary: manually update solution
      operationId: src.api.api.update_current_ea_solution
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: best solution re-evaluated
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/fitness-logbook:
    get:
      summary: get fitness logbook for this problem
      operationId: src.api.api.get_fitness_logbook
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: get fitness logbook
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /problem/{problem_id}/fitness-graph:
    get:
      summary: get fitness graph for this problem
      operationId: src.api.api.get_fitness_graph
      parameters:
        - in: path
          name: problem_id
          schema:
            type: integer
          required: true
          description: Numeric ID to get problem
      responses:
        '200':
          description: get fitness graph
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body